Fix error handling, overflow, OOB compound literals, return values

+++ include/circ_stack.h
@@ -6,7 +6,7 @@
 /*   By: pkongkha <pkongkha@student.42bangkok.com>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/13 22:27:27 by pkongkha          #+#    #+#             */
-/*   Updated: 2026/04/18 13:39:09 by pkongkha         ###   ########.fr       */
+/*   Updated: 2026/04/21 16:07:25 by pkongkha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ typedef struct s_circ_stack
 	size_t	size;
 }	t_circ_stack;
 
-size_t	circ_stack_add_strsnums(t_circ_stack *cs, char **strsnums,
+int		circ_stack_add_strsnums(t_circ_stack *cs, char **strsnums,
 			size_t nstr);
 void	circ_stack_destroy(t_circ_stack *cs);
 int		circ_stack_init(t_circ_stack *cs, size_t size);
